+++ crates/nostr/CHANGELOG.md
@@ -44,6 +44,7 @@
 - Add NIP-C7 support (https://github.com/rust-nostr/nostr/pull/1067)
 - Add NIP-60 support (https://github.com/rust-nostr/nostr/pull/1092)
 - Implement `ToBech32` trait for `Nip21`
+- Implement `From<&Event>` for `Nip19Event`
 - Add `CowTag::kind` and `CowTag::content`
 - Add `Timestamp::as_secs`
 - Add `Kind::CashuNutZapInfo` (10019) and `Kind::CashuNutZap` (9321) variants
@@ -74,6 +75,7 @@
 ### Deprecated
 
 - Deprecate `Timestamp::as_u64`
+- Deprecate `Nip19Event::from_event`
 - Deprecate `kind` field in `CommentTarget::Coordinate` variant (https://github.com/rust-nostr/nostr/pull/1035)
 
 ### Breaking changes